requirementst.txt
pyautogui
numpy 
mss
spotify
colorthief
python-dotenv
pyserial